Integrate with secure messaging
